refactor: restructured tests to match src
